move Heartbeat to separate module
